move server-related stuff into lib/server
