+++ android/adb.js
@@ -5,7 +5,7 @@ var spawn = require('win-spawn')
   , path = require('path')
   , fs = require('fs')
   , net = require('net')
-  , logger = require('../logger').get('appium')
+  , logger = require('../lib/server/logger.js').get('appium')
   , status = require('../lib/uiauto/lib/status')
   , unzipFile = require('../lib/helpers').unzipFile
   , testZipArchive = require('../lib/helpers').testZipArchive
