+++ lib/android.js
@@ -3,7 +3,7 @@
 var errors = require('./errors')
   , adb = require('../android/adb')
   , _ = require('underscore')
-  , logger = require('../logger').get('appium')
+  , logger = require('./server/logger.js').get('appium')
   , deviceCommon = require('./device')
   , status = require("./uiauto/lib/status")
   //, NotImplementedError = errors.NotImplementedError
