+++ lib/appium.js
@@ -2,9 +2,10 @@
 // https://github.com/hugs/appium/blob/master/appium/appium.py
 "use strict";
 var routing = require('./routing')
-  , logger = require('../logger').get('appium')
-  , setLogFile = require('../logger').setLogFile
-  , setWebhook = require('../logger').setWebhook
+  , loggerjs = require('./server/logger.js')
+  , logger = loggerjs.get('appium')
+  , setLogFile = loggerjs.setLogFile
+  , setWebhook = loggerjs.setWebhook
   , helpers = require('./helpers')
   , downloadFile = helpers.downloadFile
   , unzipApp = helpers.unzipApp
