+++ lib/chrome_android.js
@@ -3,7 +3,7 @@
 var Android = require('./android').Android
   , _ = require('underscore')
   , proxyTo = require('./device').proxyTo
-  , logger = require('../logger').get('appium')
+  , logger = require('./server/logger.js').get('appium')
   , exec = require('child_process').exec
   , spawn = require('child_process').spawn
   , async = require('async')
