+++ lib/controller.js
@@ -2,7 +2,7 @@
 // https://github.com/hugs/appium/blob/master/appium/server.py
 "use strict";
 var status = require('./uiauto/lib/status')
-  , logger = require('../logger.js').get('appium')
+  , logger = require('./server/logger.js').get('appium')
   , _s = require("underscore.string")
   , swig = require('swig')
   , path = require('path')
@@ -983,5 +983,6 @@ exports.guineaPig = function(req, res) {
 };
 
 exports.getTemplate = function(templateName) {
-  return swig.compileFile(path.resolve(__dirname, "templates/" + templateName + ".html"));
+  return swig.compileFile(path.resolve(__dirname, "server", "templates",
+        templateName + ".html"));
 };
