+++ lib/device.js
@@ -5,7 +5,7 @@ var errors = require('./errors')
   , _ = require('underscore')
   , exec = require('child_process').exec
   , status = require("./uiauto/lib/status")
-  , logger = require('../logger').get('appium');
+  , logger = require('./server/logger.js').get('appium');
 
 var UnknownError = errors.UnknownError
   , ProtocolError = errors.ProtocolError;
