+++ lib/firefoxos.js
@@ -2,7 +2,7 @@
 
 var errors = require('./errors')
   , _ = require('underscore')
-  , logger = require('../logger').get('appium')
+  , logger = require('./server/logger.js').get('appium')
   , net = require('net')
   , deviceCommon = require('./device')
   , status = require("./uiauto/lib/status")
