+++ lib/helpers.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var logger = require('../logger').get('appium')
+var logger = require('./server/logger.js').get('appium')
   , fs = require('fs')
   , ncp = require('ncp').ncp
   , request = require('request')
