+++ lib/hybrid/ios/remote-debugger.js
@@ -2,7 +2,7 @@
 /* DEPENDENCIES */
 
 var net = require('net')
-  , appLogger = require('../../../logger.js').get('appium')
+  , appLogger = require('../../server/logger.js').get('appium')
   , _ = require('underscore')
   , messages = require('./remote-messages.js')
   , atoms = require('./webdriver-atoms')
