+++ lib/hybrid/ios/webkit-remote-debugger.js
@@ -2,7 +2,7 @@
 /* DEPENDENCIES */
 
 var net = require('net')
-  , appLogger = require('../../../logger.js').get('appium')
+  , appLogger = require('../../server/logger.js').get('appium')
   , _ = require('underscore')
   , atoms = require('./webdriver-atoms')
   , status = require("../../uiauto/lib/status")
