+++ lib/ios.js
@@ -3,7 +3,7 @@ var path = require('path')
   , rimraf = require('rimraf')
   , fs = require('fs')
   , _ = require('underscore')
-  , logger = require('../logger').get('appium')
+  , logger = require('./server/logger.js').get('appium')
   , sock = '/tmp/instruments_sock'
   , glob = require('glob')
   , exec = require('child_process').exec
