+++ lib/proxy.js
@@ -1,7 +1,7 @@
 "use strict";
 
 var _s = require('underscore.string')
-  , logger = require('../logger.js').get('appium')
+  , logger = require('./server/logger.js').get('appium')
   , doRequest = require('./device').doRequest
   , respondError = require('./responses').respondError
   , _ = require('underscore');
