+++ lib/responses.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var logger = require('../logger').get('appium')
+var logger = require('./server/logger.js').get('appium')
   , status = require('./uiauto/lib/status')
   , _ = require('underscore');
