+++ lib/selendroid.js
@@ -8,7 +8,7 @@ var errors = require('./errors')
   , proxyTo = deviceCommon.proxyTo
   , getLog = deviceCommon.getLog
   , getLogTypes = deviceCommon.getLogTypes
-  , logger = require('../logger').get('appium')
+  , logger = require('./server/logger.js').get('appium')
   , status = require("./uiauto/lib/status")
   , fs = require('fs')
   , async = require('async')
