+++ lib/server/logger.js
