+++ lib/server/main.js
@@ -4,16 +4,16 @@ var http = require('http')
   , path = require('path')
   , fs = require('fs')
   , logger = require('./logger').get('appium')
-  , appium = require('./lib/appium')
+  , appium = require('../appium')
   , bodyParser = require('./middleware').parserWrap
-  , status = require('./lib/uiauto/lib/status')
-  , appiumVer = require('./package.json').version
+  , status = require('../uiauto/lib/status')
+  , appiumVer = require('../../package.json').version
   , appiumRev = null
   , async = require('async')
   , _ = require("underscore")
-  , parser = require('./lib/parser')
+  , parser = require('../parser')
   , io = require('socket.io')
-  , gridRegister = require('./grid_register');
+  , gridRegister = require('../../grid_register');
 
 var allowCrossDomain = function(req, res, next) {
   res.header('Access-Control-Allow-Origin', '*');
@@ -79,7 +79,7 @@ var main = function(args, readyCb, doneCb) {
 
   rest.configure(function() {
     rest.use(express.favicon());
-    rest.use(express.static(path.join(__dirname, '/app/static')));
+    rest.use(express.static(path.join(__dirname, 'static')));
     rest.use(allowCrossDomain);
     if (!args.quiet) {
       rest.use(express.logger('dev'));
@@ -100,7 +100,7 @@ var main = function(args, readyCb, doneCb) {
   appiumServer.attachTo(rest);
 
   var checkSetup = function(cb) {
-    var configFile = path.resolve(__dirname, ".appiumconfig");
+    var configFile = path.resolve(__dirname, "..", "..", ".appiumconfig");
     fs.readFile(configFile, function(err, data) {
       if (err) {
         logger.error("Could not find config file; looks like config hasn't " +
