+++ lib/server/static/favicon.ico
